Extracted Youtube fields into Fragment
